module15/nosql/requirements.txt
flask
requests
google-cloud-firestore
python-dotenv